Deleted Logo folder and moved Logo to app/components - deleted:    Logo/styles.css   moved to app/styles.css - modified:   app/components/Layout.jsx   fixed reference to Logo component - renamed:    Logo/index.js -> app/components/Logo.jsx - modified:   app/styles.css   added logo styles
