--- client/Logo/styles.css
@@ -1,14 +1,0 @@
-/* styles.css */
-
-.logo {
-  font-family: 'Amatic SC', cursive;
-  font-size: 64px;
-  color: #30cae9;
-  font-weight: bold;
-}
-
-.logo img {
-  width: 50px;
-  padding-left: 15px;
-  vertical-align: middle;
-}
